WIP MakeService attempt: MakesMeCry
First attempt at moving this all into a struct implementing the
MakeService trait. For now the organization is a bit gross, since
want to just call the original function for now to try to get this to
the point where it will compile, then will deal with restructuring
afterwards.

Have gotten to the point where the trait constraints seem to be
satisfied, however it's unclear how the hyper server Builder wants to be
given a MakeService for `serve()`, thus this is not compiling as of yet.